# internal/web/static/openapi.json
{
    "swagger": "2.0",
    "info": {
        "description": "The API the Tunnel Manager screens are built on. Every path\nbelow sits under /api and needs a session, except POST\n/api/login and GET /api/setup.\n\nLogging in sets two cookies. Over HTTPS they are named\n__Host-tm_session and __Host-tm_csrf, and over plain HTTP they\nare tm_session and tm_csrf. A browser sends them back on its\nown, so \"Try it out\" on this page works as soon as you are\nlogged in to the screens in the same browser.\n\nEvery POST, PUT and DELETE also needs the token of the session\nin an X-CSRF-Token header. POST /api/login answers with it in\ndata.csrf_token; paste it into Authorize above and this page\nsends it. A GET needs no token.\n\nEvery answer has the same shape: {\"success\":true,\"data\":...}\nor {\"success\":false,\"error\":\"...\"}. An answer that says no\ncarries error_code, which names the refusal and is what a\nscript should match on, and error_args, the values its English\nsentence was written with.",
        "title": "Tunnel Manager API",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "3.8.3"
    },
    "basePath": "/api",
    "paths": {
        "/account": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "What the account is called",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.accountView"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/api.errorBody"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "CSRFToken": []
                    }
                ],
                "description": "Takes current_password and username, new_password or both, changes them and signs out every other session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Change the username, the password or both",
                "parameters": [
                    {
                        "description": "The current password, and what to change",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.accountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.accountChanged"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Nothing was asked for, or the new password is refused",
                        "schema": {
                            "$ref": "#/definitions/api.errorBody"
                        }
                    },
                    "401": {
                        "description": "The current password is wrong",
                        "schema": {
                            "$ref": "#/definitions/api.errorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.accountChanged": {
            "type": "object",
            "properties": {
                "password_changed": {
                    "type": "boolean"
                },
                "sessions_ended": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                },
                "username_changed": {
                    "type": "boolean"
                }
            }
        },
        "api.accountRequest": {
            "type": "object",
            "properties": {
                "current_password": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.accountView": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                }
            }
        },
...
        "api.errorArgs": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "api.errorBody": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "error_args": {
                    "$ref": "#/definitions/api.errorArgs"
                },
                "error_code": {
                    "$ref": "#/definitions/api.errorCode"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "api.errorCode": {
            "type": "string",
            "enum": [
                "unspecified",
                "request.body.invalid",
                "request.validation.failed",
                "db.transaction.begin_failed",
                "db.transaction.commit_failed",
                "list.page.not_a_number",
                "list.size.unsupported",
                "auth.session.required",
                "auth.csrf.refused",
                "auth.credentials.invalid",
                "auth.session.create_failed",
                "auth.setup.required",
                "auth.setup.already_done",
                "auth.setup.failed",
                "auth.password.too_short",
                "auth.password.too_long",
                "auth.attempts.too_many",
                "auth.password_attempts.too_many",
                "account.read_failed",
                "account.store_failed",
                "account.password.hash_failed",
                "account.username.empty",
                "account.request.invalid",
                "account.nothing_to_change",
                "account.current_password.wrong",
                "account.username.unchanged",
                "account.new_password.unchanged",
                "account.new_password.too_short",
                "account.new_password.too_long",
                "host.id.invalid",
                "host.not_found",
                "host.fetch_failed",
                "host.list_failed",
                "host.create_failed",
                "host.update_failed",
                "host.delete_failed",
                "host.password.encrypt_failed",
                "host.private_key.encrypt_failed",
                "host.create.no_login",
                "host.create.key_refused",
                "host.create.service_ports_read_failed",
                "host.create.service_ports_store_failed",
                "host.update.key_refused",
                "host.update.passphrase_without_key",
                "host.host_key.nothing_to_approve",
                "host.host_key.fingerprint_changed",
                "host.host_key.password_wrong",
                "service_port.id.invalid",
                "service_port.not_found",
                "service_port.fetch_failed",
                "service_port.list_failed",
                "service_port.create_failed",
                "service_port.update_failed",
                "service_port.delete_failed",
                "service_port.create.hosts_read_failed",
                "service_port.create.hosts_store_failed",
                "assignment.add_and_remove",
                "assignment.service_port.not_found",
                "assignment.update_failed",
                "status.desired_count_failed",
                "status.fetch_failed",
                "settings.read_failed",
                "settings.store_failed",
                "settings.refused",
                "settings.ui_language.unsupported",
                "certificate.https_off",
                "certificate.served.read_failed",
                "certificate.renew_failed",
                "certificate.install.refused",
                "certificate.store_failed",
                "certificate.read_back_failed",
                "logs.lines.not_a_number",
                "logs.lines.below_one",
                "logs.file.not_configured",
                "logs.file.missing",
                "logs.file.read_failed",
                "logs.clear.password_missing",
                "logs.clear.password_wrong",
                "logs.clear.failed",
                "update.check_failed",
                "update.not_installable",
                "update.password_missing",
                "update.password_wrong",
                "update.install_failed",
                "export.password.required",
                "export.password.too_short",
                "export.password.too_long",
                "export.hosts.read_failed",
                "export.service_ports.read_failed",
                "export.assignments.read_failed",
                "export.host.secrets_unreadable",
                "export.seal_failed",
                "import.file.missing",
                "import.password.required",
                "import.file.not_an_export",
                "import.file.damaged",
                "import.password.wrong",
                "import.file.open_failed",
                "import.file.not_ours",
                "import.file.wrong_kind",
                "import.file.newer_format",
                "import.file.newer_format_by",
                "import.file.empty",
                "import.tunnels.unreadable",
                "import.host.refused",
                "import.host.no_login",
                "import.hosts.read_failed",
                "import.host.password_encrypt_failed",
                "import.host.key_refused",
                "import.host.private_key_encrypt_failed",
                "import.host.replace_failed",
                "import.host.create_failed",
                "import.service_port.refused",
                "import.service_ports.read_failed",
                "import.service_port.create_failed",
                "import.service_port.two_rows",
                "import.service_port.replace_failed",
                "import.assignments.host_read_failed",
                "import.assignments.clear_failed",
                "import.assignments.store_failed",
                "import.settings.unreadable",
                "import.settings.refused",
                "uninstall.password.wrong"
            ],
            "x-enum-varnames": [
                "errorCodeUnspecified",
                "errRequestBodyInvalid",
                "errRequestValidationFailed",
                "errTransactionBeginFailed",
                "errTransactionCommitFailed",
                "errListPageNotANumber",
                "errListSizeUnsupported",
                "errAuthRequired",
                "errAuthCSRFRefused",
                "errAuthCredentialsInvalid",
                "errAuthSessionCreateFailed",
                "errAuthSetupRequired",
                "errAuthSetupAlreadyDone",
                "errAuthSetupFailed",
                "errAuthPasswordTooShort",
                "errAuthPasswordTooLong",
                "errAuthTooManyAttempts",
                "errAuthPasswordTooManyAttempts",
                "errAccountReadFailed",
                "errAccountStoreFailed",
                "errAccountHashFailed",
                "errAccountUsernameEmpty",
                "errAccountRequestInvalid",
                "errAccountNothingToChange",
                "errAccountPasswordWrong",
                "errAccountUsernameUnchanged",
                "errAccountPasswordUnchanged",
                "errAccountNewPasswordShort",
                "errAccountNewPasswordLong",
                "errHostIDInvalid",
                "errHostNotFound",
                "errHostFetchFailed",
                "errHostListFailed",
                "errHostCreateFailed",
                "errHostUpdateFailed",
                "errHostDeleteFailed",
                "errHostPasswordEncryptFailed",
                "errHostKeyEncryptFailed",
                "errHostCreateNoLogin",
                "errHostCreateKeyRefused",
                "errHostCreateServicePortsRead",
                "errHostCreateServicePortsStore",
                "errHostUpdateKeyRefused",
                "errHostUpdatePassphraseAlone",
                "errHostKeyNothingToApprove",
                "errHostKeyFingerprintChanged",
                "errHostKeyPasswordWrong",
                "errServicePortIDInvalid",
                "errServicePortNotFound",
                "errServicePortFetchFailed",
                "errServicePortListFailed",
                "errServicePortCreateFailed",
                "errServicePortUpdateFailed",
                "errServicePortDeleteFailed",
                "errServicePortCreateHostsRead",
                "errServicePortCreateHostsStore",
                "errAssignmentAddAndRemove",
                "errAssignmentServicePortsMissing",
                "errAssignmentUpdateFailed",
                "errStatusDesiredCountFailed",
                "errStatusFetchFailed",
                "errSettingsReadFailed",
                "errSettingsStoreFailed",
                "errSettingsRefused",
                "errSettingsLanguageUnsupported",
                "errCertificateHTTPSOff",
                "errCertificateServedUnread",
                "errCertificateRenewFailed",
                "errCertificateInstallRefused",
                "errCertificateStoreFailed",
                "errCertificateReadBackFailed",
                "errLogsLinesNotANumber",
                "errLogsLinesBelowOne",
                "errLogsFileNotConfigured",
                "errLogsFileMissing",
                "errLogsFileReadFailed",
                "errLogsClearPasswordMissing",
                "errLogsClearPasswordWrong",
                "errLogsClearFailed",
                "errUpdateCheckFailed",
                "errUpdateNotInstallable",
                "errUpdatePasswordMissing",
                "errUpdatePasswordWrong",
                "errUpdateInstallFailed",
                "errExportPasswordRequired",
                "errExportPasswordTooShort",
                "errExportPasswordTooLong",
                "errExportHostsReadFailed",
                "errExportServicePortsRead",
                "errExportAssignmentsRead",
                "errExportHostSecretsSealed",
                "errExportSealFailed",
                "errImportFileMissing",
                "errImportPasswordRequired",
                "errImportFileNotAnExport",
                "errImportFileDamaged",
                "errImportPasswordWrong",
                "errImportFileOpenFailed",
                "errImportFileNotOurs",
                "errImportFileWrongKind",
                "errImportFileNewerFormat",
                "errImportFileNewerFormatBy",
                "errImportFileEmpty",
                "errImportTunnelsUnreadable",
                "errImportHostRefused",
                "errImportHostNoLogin",
                "errImportHostsReadFailed",
                "errImportHostPasswordEncrypt",
                "errImportHostKeyRefused",
                "errImportHostKeyEncrypt",
                "errImportHostReplaceFailed",
                "errImportHostCreateFailed",
                "errImportServicePortRefused",
                "errImportServicePortsReadFailed",
                "errImportServicePortCreate",
                "errImportServicePortTwoRows",
                "errImportServicePortReplace",
                "errImportAssignmentsHostRead",
                "errImportAssignmentsClearFailed",
                "errImportAssignmentsStoreFailed",
                "errImportSettingsUnreadable",
                "errImportSettingsRefused",
                "errUninstallPasswordWrong"
            ]
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CSRFToken": {
            "description": "The token POST /api/login answers with, in\ndata.csrf_token. Required on every POST, PUT\nand DELETE. The session cookie goes with it and\nis sent by the browser on its own.",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    }
}